Trim build.sh and remove unused Lua files
CirnOS was made with the belief that learning everything about
the Raspberry Pi's internals could take months or years. Through
looking at components of other operating systems that and drivers
that worked, CirnOS was created as an amalgamation of various
working pieces from different systems.

Now that all of the basic functionality is in place, the remaining
features of these different systems that CirnOS does not use can
be removed from the project files. The various unnecessary
command line arguments in build.sh have been removed so that every
one now has a purpose.

Files lua.hpp and luac.c have also been removed as they did not
serve any purpose in the complete CirnOS project.