Added hardware PWM wrappers
The Raspberry Pi original models and zero model both
only have a single pin with hardware PWM enabled --
pin 18. This pin is automatically used in the PWM
functions added to Lua to avoid redundancy.

Support for software PWM should eventually be added
to all of the usable ports in due time. The ability
to change the speed of the PWM should also be
configurable.